+++ python/lsst/ts/ess/csc/data_client/siglent_ssa3000x_spectrum_analyzer_data_client.py
@@ -131,7 +131,9 @@ additionalProperties: false
         await self.disconnect()
 
         if self.simulation_mode == 0:
-            self.client = tcpip.Client(host=self.config.host, port=self.config.port)
+            self.client = tcpip.Client(
+                host=self.config.host, port=self.config.port, log=self.log
+            )
             await asyncio.wait_for(self.client.start_task, self.config.connect_timeout)
         else:
             self.log.info(
@@ -177,7 +179,7 @@ additionalProperties: false
                     await self.write(self.query_trace_data_cmd)
                     assert self.client is not None  # make mypy happy
                     read_bytes = await asyncio.wait_for(
-                        self.client.read_until(TERMINATOR),
+                        self.client.readuntil(TERMINATOR),
                         timeout=self.config.read_timeout,
                     )
                     raw_data = read_bytes.decode().strip()